+++ app/src/main/java/es/voghdev/hellocompose/components.kt
@@ -1,0 +1,167 @@
+package es.voghdev.hellocompose
+
+import HelloAppNavigation
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.Layout
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import androidx.navigation.compose.rememberNavController
+import es.voghdev.hellocompose.ui.theme.HelloComposeTheme
+
+private val topics = listOf(
+    "Arts & Crafts", "Beauty", "Books", "Business", "Comics", "Culinary",
+    "Design", "Fashion", "Film", "History", "Maths", "Music", "People", "Philosophy",
+    "Religion", "Social sciences", "Technology", "TV", "Writing"
+)
+
+@Composable
+fun HelloComposeApp() {
+    val navController = rememberNavController()
+    HelloAppNavigation(navController)
+    HelloComposeTheme {
+        MainScreen(
+            onOpenGridClick = {
+                navController.navigate("chips")
+            },
+            onOpenDraggableClick = {
+                navController.navigate("items")
+            }
+        )
+    }
+}
+
+@Composable
+fun MainScreen(
+    onOpenGridClick: () -> Unit,
+    onOpenDraggableClick: () -> Unit
+) {
+    Surface(color = MaterialTheme.colors.background) {
+        Column {
+            Button(onClick = onOpenGridClick) {
+                Text("Item chips example")
+            }
+            Button(onClick = onOpenDraggableClick) {
+                Text("Draggable example")
+            }
+        }
+    }
+}
+
+@Composable
+fun ChipsScreen() {
+    Surface(color = MaterialTheme.colors.background) {
+        BodyContent()
+    }
+}
+
+@Composable
+fun BodyContent(modifier: Modifier = Modifier) {
+    StaggeredGrid(modifier = modifier) {
+        for (topic in topics) {
+            Chip(modifier = Modifier.padding(8.dp), text = topic)
+        }
+    }
+}
+
+@Preview
+@Composable
+fun LayoutsCodelabPreview() {
+    HelloComposeTheme {
+        BodyContent()
+    }
+}
+
+@Composable
+fun StaggeredGrid(
+    modifier: Modifier = Modifier,
+    rows: Int = 3,
+    content: @Composable () -> Unit
+) {
+    Layout(
+        modifier = modifier,
+        content = content
+    ) { measurables, constraints ->
+        // Keep track of the width of each row
+        val rowWidths = IntArray(rows) { 0 }
+
+        // Keep track of the max height of each row
+        val rowHeights = IntArray(rows) { 0 }
+
+        // Don't constrain child views further, measure them with given constraints
+        // List of measured children
+        val placeables = measurables.mapIndexed { index, measurable ->
+
+            // Measure each child
+            val placeable = measurable.measure(constraints)
+
+            // Track the width and max height of each row
+            val row = index % rows
+            rowWidths[row] += placeable.width
+            rowHeights[row] = Math.max(rowHeights[row], placeable.height)
+
+            placeable
+        }
+
+        // Grid's width is the widest row
+        val width = rowWidths.maxOrNull()
+            ?.coerceIn(constraints.minWidth.rangeTo(constraints.maxWidth))
+            ?: constraints.minWidth
+
+        // Grid's height is the sum of the tallest element of each row
+        // coerced to the height constraints
+        val height = rowHeights.sumOf { it }
+            .coerceIn(constraints.minHeight.rangeTo(constraints.maxHeight))
+
+        // Y of each row, based on the height accumulation of previous rows
+        val rowY = IntArray(rows) { 0 }
+        for (i in 1 until rows) {
+            rowY[i] = rowY[i - 1] + rowHeights[i - 1]
+        }
+
+        // Set the size of the parent layout
+        layout(width, height) {
+            // x cord we have placed up to, per row
+            val rowX = IntArray(rows) { 0 }
+
+            placeables.forEachIndexed { index, placeable ->
+                val row = index % rows
+                placeable.placeRelative(
+                    x = rowX[row],
+                    y = rowY[row]
+                )
+                rowX[row] += placeable.width
+            }
+        }
+    }
+}
+
+@Composable
+fun Chip(modifier: Modifier = Modifier, text: String) {
+    Card(
+        modifier = modifier,
+        border = BorderStroke(color = Color.Black, width = Dp.Hairline),
+        shape = RoundedCornerShape(8.dp)
+    ) {
+        Row(
+            modifier = Modifier.padding(start = 8.dp, top = 4.dp, end = 8.dp, bottom = 4.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Box(
+                modifier = Modifier
+                    .size(16.dp, 16.dp)
+                    .background(color = MaterialTheme.colors.secondary)
+            )
+            Spacer(Modifier.width(4.dp))
+            Text(text = text)
+        }
+    }
+}
